+++ Assets/Code/GiantsAttack/SubStageExecutorSpecialAttack.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using GameCore.UI;
+using SleepDev.Vibration;
 
 namespace GiantsAttack
 {
@@ -44,6 +45,8 @@ namespace GiantsAttack
             _target.OnAttack();
             MinusTarget();
             _ui.Flash.Play();
+            VibrationManager.VibrManager.PlaySimple();
+            PrintEvent();
         }
 
         private void OnCompleted()