+++ Assets/MadPixel/MAXHelper/AdsManager.cs
@@ -3,16 +3,14 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Events;
 using UnityEngine.Networking;
-#if MADPIXEL_AMAZON_DROID && !UNITY_EDITOR
-using AmazonAds;
-#endif
+
 
 namespace MAXHelper {
     
     [RequireComponent(typeof(TermsAndATT))]
     [RequireComponent(typeof(AppLovinComp))]
     public class AdsManager : MonoBehaviour {
-        private const string version = "1.2.8";
+        private const string version = "1.2.9";
         public enum EResultCode {OK = 0, NOT_LOADED, ADS_FREE, ON_COOLDOWN, ERROR}
         public enum EAdType {REWARDED, INTER, BANNER}
 
@@ -435,6 +433,26 @@ namespace MAXHelper {
             }
             return (false);
         }
+
+        /// <summary>
+        /// Returns mandatory Cooldown between interstitials, if set
+        /// </summary>
+        public static int GetCooldownBetweenInters() {
+            if (Exist) {
+                return Instance.CooldownBetweenInterstitials;
+            }
+
+            return 0;
+        }
+
+        /// <summary>
+        /// Restarts interstitial cooldown (it already restarts automatically after an ad is watched)
+        /// </summary>
+        public static void RestartInterstitialCooldown() {
+            if (Exist) {
+                Instance.RestartInterCooldown();
+            }
+        }
         #endregion
 
         #region Helpers
@@ -445,9 +463,6 @@ namespace MAXHelper {
         }
         
         private void InitApplovinInternal() {
-#if MADPIXEL_AMAZON_DROID && !UNITY_EDITOR
-            InitAmazon();
-#endif
             LastInterShown = -CooldownBetweenInterstitials;
 
             AppLovin.Init(CustomSettings);
@@ -468,29 +483,6 @@ namespace MAXHelper {
             AdsInstigatorObj = objectRef;
             CallbackPending = Callback;
         }
-
-#if MADPIXEL_AMAZON_DROID && !UNITY_EDITOR
-        public void InitAmazon() {
-#if UNITY_ANDROID
-            string amazonSDKKey = CustomSettings.AmazonSDKKey;
-#else
-            string amazonSDKKey = CustomSettings.AmazonSDKKey_IOS;
-#endif
-
-            if (!string.IsNullOrEmpty(amazonSDKKey)) {
-                Amazon.Initialize(amazonSDKKey);
-                Debug.Log($"[MadPixel] Amazon Init");
-
-                Amazon.EnableLogging(false);
-                Amazon.EnableTesting(false);
-
-                Amazon.SetAdNetworkInfo(new AdNetworkInfo(DTBAdNetwork.MAX));
-            }
-            else {
-                Debug.LogError($"[MadPixel] Amazon cant init! SDK is empty!");
-            }
-        }
-#endif
 
         private void ShowAdInner(EAdType AdType, string Placement) {
             CurrentAdInfo = new AdInfo(Placement, AdType);