+++ Assets/MadPixel/MAXHelper/Editor/MAXHelperDefineSymbols.cs
@@ -28,6 +28,7 @@ namespace MAXHelper {
                 }
             }
 
+
             PlayerSettings.SetScriptingDefineSymbols(NamedBuildTarget.Android, string.Join(";", scriptingDefinesStringList.ToArray()));
         }