+++ Assets/MadPixel/MAXHelper/Editor/MAXHelperInitWindow.cs
@@ -13,7 +13,8 @@ namespace MAXHelper {
         private const string AMAZON_PACKAGE_PATH = "Assets/MadPixel/MAXHelper/Configs/Amazon_APS.unitypackage";
         private const string MEDIATIONS_PATH = "Assets/MAXSdk/Mediation/";
         private const string AMAZON_PATH = "Assets/Amazon/Plugins";
-        private const string MPC_FOLDER = "https://drive.google.com/drive/u/0/folders/1Mo36yT_dWR36lZvRWmkLHArtxna1zzS2";
+        private const string MPC_FOLDER = "https://github.com/MadPixelDevelopment/MadPixelCore/releases";
+        private const string MAX_PACK_INDEPENDENT = "https://github.com/MadPixelDevelopment/MadPixelCore/raw/main/Assets/MadPixel/MAXHelper/Configs/MaximumPack.unitypackage";
 
         private const string ADS_DOC =
             "https://docs.google.com/document/d/1lx9wWCD4s8v4aXH1pb0oQENz01UszdalHtnznmQv2vc/edit#heading=h.y039lv8byi2i";
@@ -104,8 +105,6 @@ namespace MAXHelper {
                     DrawTestPart();
 
                     DrawInstallButtons();
-
-                    DrawAmazon();
 
                     DrawLinks();
                 }
@@ -243,17 +242,32 @@ namespace MAXHelper {
             using (new EditorGUILayout.VerticalScope("box")) {
                 GUILayout.BeginHorizontal();
                 GUILayout.Space(10);
-                GUI.enabled = false;
-                if (GUILayout.Button(new GUIContent("Minimum pack is installed"), buttonFieldWidth)) {
-                    // nothing here
+
+                if (!MackPackUnitypackageExists()) {
+                    EditorGUILayout.LabelField("You dont have MaximunPack.unitypackage in your project. Probably your git added it to gitignore", sdkKeyTextFieldWidthOption);
+                    
+                    GUILayout.EndHorizontal();
+                    GUILayout.BeginHorizontal();
+                    GUILayout.Space(10);
+
+                    if (GUILayout.Button(new GUIContent("Download latest Maximum mediations package"), adMobUnitTextWidthOption)) {
+                        Application.OpenURL(MAX_PACK_INDEPENDENT);
+                    }
+
+                    GUILayout.EndHorizontal();
+                    GUILayout.Space(10);
+                    GUILayout.BeginHorizontal();
+                    GUILayout.Space(10);
                 }
 
-                GUI.enabled = true;
-
-                GUILayout.Space(5);
-                GUILayout.Space(10);
-                GUI.enabled = !bMaxVariantInstalled;
-                if (GUILayout.Button(new GUIContent(bMaxVariantInstalled ? "Maximum pack is installed" : "Install maximum pack"), buttonFieldWidth)) {
+                GUI.enabled = MackPackUnitypackageExists();
+                if (bMaxVariantInstalled) {
+                    EditorGUILayout.LabelField("You have installed default Maximum pack of mediations", sdkKeyTextFieldWidthOption);
+                    GUILayout.EndHorizontal();
+                    GUILayout.BeginHorizontal();
+                    GUILayout.Space(10);
+                }
+                if (GUILayout.Button(new GUIContent(bMaxVariantInstalled ? "Reimport maximum pack" : "Install maximum pack"), buttonFieldWidth)) {
                     AssetDatabase.ImportPackage(PACKAGE_PATH, true);
                     CheckMaxVersion();
                 }
@@ -319,90 +333,6 @@ namespace MAXHelper {
             return text;
         }
 
-        private void DrawAmazon() {
-            GUILayout.Space(16);
-            EditorGUILayout.LabelField("5. Amazon", titleLabelStyle);
-            using (new EditorGUILayout.VerticalScope("box")) {
-                GUILayout.BeginHorizontal();
-                GUILayout.Space(4);
-                bool bHasFolder = Directory.Exists(AMAZON_PATH);
-
-                if (bHasFolder) {
-                    EditorGUILayout.LabelField("Amazon plugin is installed", adMobUnitTextWidthOption);
-                    bUseAmazon = MAXHelperDefineSymbols.HasAmazonActivated();
-                    GUILayout.EndHorizontal();
-
-                    GUILayout.BeginHorizontal();
-                    GUILayout.Space(4);
-                    EditorGUILayout.LabelField(
-                        bUseAmazon ? "Amazon is used by AdsManager" : "You dont have Amazon initialized",
-                        adMobUnitTextWidthOption);
-                    GUILayout.EndHorizontal();
-
-                    GUILayout.BeginHorizontal();
-                    GUILayout.Space(4);
-                    if (GUILayout.Button(new GUIContent(bUseAmazon ? "Deactivate Amazon" : "Activate Amazon"),
-                            adMobUnitTextWidthOption)) {
-                        OnActivateAmazonClick();
-                    }
-
-                    if (bUseAmazon) {
-                        DrawAmazonValues();
-                    }
-                }
-                else {
-                    EditorGUILayout.LabelField("You dont have Amazon Plugin installed", adMobUnitTextWidthOption);
-                    GUILayout.EndHorizontal();
-
-                    GUILayout.BeginHorizontal();
-                    GUILayout.Space(4);
-                    if (GUILayout.Button(new GUIContent("Install Amazon plugin"), adMobUnitTextWidthOption)) {
-                        OnInstallAmazonPluginClick();
-                    }
-
-                    if (bUseAmazon) {
-                        GUILayout.EndHorizontal();
-
-                        GUILayout.BeginHorizontal();
-                        GUILayout.Space(4);
-                        if (GUILayout.Button(new GUIContent(bUseAmazon ? "Deactivate Amazon" : "Activate Amazon"),
-                                adMobUnitTextWidthOption)) {
-                            OnActivateAmazonClick();
-                        }
-                    }
-                }
-
-                GUILayout.EndHorizontal();
-            }
-        }
-
-        private void DrawAmazonValues() {
-            GUILayout.EndHorizontal();
-
-            GUILayout.BeginHorizontal();
-            GUILayout.Space(4);
-            CustomSettings.AmazonSDKKey = DrawTextField("Amazon SDK key (Android)", CustomSettings.AmazonSDKKey, buttonFieldWidth, adMobUnitTextWidthOption);
-            CustomSettings.AmazonSDKKey_IOS = DrawTextField("Amazon SDK key (iOS)", CustomSettings.AmazonSDKKey_IOS, buttonFieldWidth, adMobUnitTextWidthOption);
-            GUILayout.EndHorizontal();
-
-            GUILayout.BeginHorizontal();
-            GUILayout.Space(4);
-            CustomSettings.AmazonRewardedID = DrawTextField("Rewarded Ad Unit (Android)", CustomSettings.AmazonRewardedID, buttonFieldWidth, adMobUnitTextWidthOption);
-            CustomSettings.AmazonRewardedID_IOS = DrawTextField("Rewarded Ad Unit (iOS)", CustomSettings.AmazonRewardedID_IOS, buttonFieldWidth, adMobUnitTextWidthOption);
-            GUILayout.EndHorizontal();
-
-            GUILayout.BeginHorizontal();
-            GUILayout.Space(4);
-            CustomSettings.AmazonInterstitialID = DrawTextField("Interstitial Ad Unit (Android)", CustomSettings.AmazonInterstitialID, buttonFieldWidth, adMobUnitTextWidthOption);
-            CustomSettings.AmazonInterstitialID_IOS = DrawTextField("Interstitial Ad Unit (iOS)", CustomSettings.AmazonInterstitialID_IOS, buttonFieldWidth, adMobUnitTextWidthOption);
-            GUILayout.EndHorizontal();
-
-            GUILayout.BeginHorizontal();
-            GUILayout.Space(4);
-            CustomSettings.AmazonBannerID = DrawTextField("Banner Ad Unit (Android)", CustomSettings.AmazonBannerID, buttonFieldWidth, adMobUnitTextWidthOption);
-            CustomSettings.AmazonBannerID_IOS = DrawTextField("Banner Ad Unit (iOS)", CustomSettings.AmazonBannerID_IOS, buttonFieldWidth, adMobUnitTextWidthOption);
-        }
-
         #endregion
 
         #region Helpers
@@ -439,7 +369,8 @@ namespace MAXHelper {
                 return "--";
             }
 
-            versionText = versionText.Substring(10, 4);
+            int subLength = versionText.IndexOf('-');
+            versionText = versionText.Substring(10, subLength - 10);
             return versionText;
         }
 
@@ -470,6 +401,10 @@ namespace MAXHelper {
             else {
                 MAXHelperDefineSymbols.DefineSymbols(false);
             }
+        }
+
+        private bool MackPackUnitypackageExists() {
+            return File.Exists(PACKAGE_PATH);
         }
 
         #endregion