+++ Assets/MadPixel/AnalyticsHelper/Changelog.md
@@ -1,4 +1,15 @@
-﻿# Version 1.0.6 - November 28, 2023
+﻿# Version 1.0.9 - March 08, 2024
+* Analytics version 1.0.9
+
+# Version 1.0.8 - February 20, 2024
+* AppsFlyer SDK updated to 6.13.0
+* AppsFlyer AdRevenue updated to 6.9.42
+* AppsFlyer: enable TCF data collection
+
+# Version 1.0.7 - January 29, 2024
+* af_purchase sends revenue via InvariantCulture now
+
+# Version 1.0.6 - November 28, 2023
 * bUseAutoInit parameter is used to init Analytics on Applovin initialized callback
 
 # Version 1.0.5 - October 2, 2023